summary: display more useful base image names
In cases where a base image has more than one layer:

- 1a1a1a1
- 2222222 : tagged here as common-base

then all images built with it will have an untagged layer[0], in which
case the summary can make it hard to identify what base image is
actually being used.

This commit adds some logic to build up the tree of layers, and if a
layer has some unique descendant that is tagged, we will display that
tag as the "base image name".

If there is no unique tagged descendant, we still use ?. This covers
most cases, but future work might want to find which path we are on and
find the unique earliest tag there

Signed-off-by: Michael McCracken <mikmccra@cisco.com>